requirements.txt
fastapi
sqlalchemy
python-jose[cryptography]
httpx
uvicorn
pydantic
pytest-mock
allure-pytest

pytest
psycopg2-binary
pandas
pyarrow
awscli==1.37.0
boto3==1.15.18
botocore==1.36.0
s3transfer==0.11.0
requests
python-dotenv